feat: [trace-drill-down-chart-graph-consolidation-004] - [Add focused regression evidence for cross-surface graph visual consistency]
