+++ docs/internal/processes/development-guide-relevant-files.md
@@ -150,7 +150,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `ui/src/features/header/dashboard-export-dialog.tsx` | Dashboard export host seam | Dashboard-owned export overlays should stay as thin host components that bridge the dialog open-store, current-factory export query, and shared `ExportFactoryDialog` surface rather than moving fetch or mutation ownership into header buttons. |
 | `ui/src/features/export/export-factory-dialog.tsx` | Shared export-dialog reset seam | Async current-factory refreshes may update the seeded factory name while the dialog is already open; reset the form on open and close, but preserve a user-edited export name unless the input is still mirroring the previous seed so background refreshes do not clobber an in-progress export filename. |
 | `ui/src/features/workflow-activity/current-activity-import-controller.ts` | Shared workflow-import controller seam | Workflow graph drag-drop, preview, and activation state should stay reusable through this controller so dashboard-owned dialogs can reuse the same import flow without remounting feature-local widget state. |
-| `ui/src/App.stories.tsx` | Dashboard Storybook smoke surface | Current-selection and dashboard workflow stories should assert observable headings, dispatch markers, mixed inference/script selected-work history output, pending or unavailable inference copy, localized dashboard-header states, and prompt-editor help or validation flows here so browser-backed regression coverage follows the customer-visible contract instead of helper structure. |
+| `ui/src/App.stories.tsx` and `ui/src/stories/*.tsx` | Dashboard Storybook smoke surface | Current-selection and dashboard workflow stories should assert observable headings, dispatch markers, mixed inference/script selected-work history output, pending or unavailable inference copy, localized dashboard-header states, and prompt-editor help or validation flows here so browser-backed regression coverage follows the customer-visible contract instead of helper structure; when `App.stories.tsx` approaches the repo file-length guard, move Storybook-only harness components into `ui/src/stories/` sidecars instead of suppressing the lint rule on the main smoke file. |
 | `ui/src/App.test.tsx` | Dashboard app-level regression surface | Keep broad dashboard-shell, export/import, timeline, submit-work, and trace fallback regressions here; when current-selection flows grow beyond the repo lint thresholds, move that coverage into a dedicated app-level seam instead of suppressing Biome guards. |
 | `ui/src/features/submit-work/use-submit-work-widget.ts`, `ui/src/features/submit-work/submit-work-card.tsx`, `ui/src/features/submit-work/submit-work-card.stories.tsx`, `ui/src/api/work/api.test.ts`, and `ui/src/api/work/api.additional.test.ts` | Dashboard submit-work mutation seam | Submit-work field rules, ready/guidance status copy, post-success draft clearing, and the stable action-row layout contract should stay normalized across the widget hook, card, and feature story here: the hook owns trim-aware validation plus draft reset behavior, the card owns accessible field wiring and inline validation, configured work-type selection stays preserved only while still valid, unary submit must require a non-blank request name while keeping request text optional, and browser-backed story regressions should prove the submit button keeps the same measured right edge across ready, submitting, success, error, and wrapped validation states. Wrapper regressions for explicit empty payloads or structured errors should assert the generated camelCase OpenAPI contract (`workTypeName`, `traceId`) instead of ad hoc snake_case request or response shapes. |
 | `ui/src/features/timeline/state/timeline/replayWorldState.ts` | Timeline replay reducer seam | Live timeline replay should reconstruct active and completed dispatch state here by reading canonical event context first and falling back to supported legacy payload fields only at the request and response consumption sites. |
@@ -238,6 +238,7 @@ This inventory records the checked-in files and directories that the maintainer 
 - Keep frontend V8 coverage thresholds in `ui/vite.config.ts` scoped to app/runtime source; exclude replay fixture files, CSS, barrel exports, and Node tooling scripts from the denominator while preserving separate Vitest coverage for script behavior.
 - When dashboard UI coverage must prove a shared button treatment, prefer Storybook assertions over rendered computed styles against an existing customer-visible dashboard action in `ui/src/App.stories.tsx` instead of adding test-only DOM metadata to shared primitives.
 - When dashboard locale-switcher behavior changes, update both the focused header coverage in `ui/src/features/header/dashboard-header.test.tsx` and the shell-level locale propagation checks in `ui/src/App.test.tsx` and `ui/src/App.stories.tsx` so the compact control contract and cross-surface localization flow stay aligned.
+- When Storybook-only dashboard harness controls need local labels outside the product localization path, keep the harness component in `ui/src/stories/` and mark the exact diagnostic-only JSX text with `hardcoded-ui-copy-exception: non-product-diagnostic` so the hardcoded-copy guard does not treat test controls as shippable UI copy.
 - When submit-work success behavior changes in the dashboard shell, update the browser-backed `DashboardSubmitWorkIntegrationSmoke` story in `ui/src/App.stories.tsx` alongside the feature hook or card so post-submit field resets and button-enablement rules stay aligned with the real widget contract.
 - When submit-work layout changes in the dashboard shell, keep the browser-backed geometry assertion in `ui/src/features/submit-work/submit-work-card.stories.tsx` aligned with the action row so state-copy changes cannot reintroduce horizontal button drift.
 - When dashboard UI work needs reusable controls or states, add or extend primitives under `ui/src/components/ui/` first and keep feature folders focused on composition, data shaping, and product-specific behavior rather than bespoke chrome.
